:bug: does fixes amount when transform to int, sometimes removes one cent

+++ view/frontend/web/js/core/checkout/PaymentMethodController.js
@@ -505,7 +505,7 @@ PaymentMethodController.prototype.updateTotal = function(interest, grandTotal, s
 
     /**@fixme Move gettotals() to PlatformFormBiding */
     var total = paymentMethodController.platformConfig.updateTotals.getTotals()();
-    interest = (parseInt(interest * 100)) / 100;
+    interest = (parseFloat(interest * 100)) / 100;
 
     if (interest < 0) {
         interest = 0;
@@ -516,7 +516,7 @@ PaymentMethodController.prototype.updateTotal = function(interest, grandTotal, s
 
     for (var i = 0, len = total.total_segments.length; i < len; i++) {
         if (total.total_segments[i].code === "grand_total") {
-            total.total_segments[i].value = (parseInt(grandTotal * 100)) / 100;
+            total.total_segments[i].value = (parseFloat(grandTotal * 100)) / 100;
             continue;
         }
         if (total.total_segments[i].code === "tax") {
